implement new convergence criterion for pele

+++ website/minimization/lj38/pele-lbfgs-M4-maxstep0.2/gmin_getdata.py
@@ -22,7 +22,8 @@ contributor = "Jacob Stevenson"
 
 if len(sys.argv) > 1:
     datafile_orig = sys.argv[1]
-    if os.path.isfile(datafile_orig):
+    datafile = datafile_orig
+    if False and os.path.isfile(datafile_orig):
         if datafile_orig != datafile:
             shutil.copy2(datafile_orig, datafile)
 if len(sys.argv) > 2:
